Update file path.
