+++ src/main/java/shef/data/User.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.sps.data;
+package shef.data;
 
 import java.util.ArrayList;
