+++ src/main/java/shef/servlets/ProfilePicUploadUrlServlet.java
@@ -12,28 +12,26 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.sps.servlets;
+package shef.servlets;
 
-import com.google.appengine.api.users.UserService;
-import com.google.appengine.api.users.UserServiceFactory;
+import com.google.appengine.api.blobstore.BlobstoreService;
+import com.google.appengine.api.blobstore.BlobstoreServiceFactory;
 import java.io.IOException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-/** Servlet that returns a sign-up URL for the sign-up page. */
-@WebServlet("/sign-up")
-public class SignUpServlet extends HttpServlet {
+/** Servlet that returns the Blobstore upload url for profile pictures. */
+@WebServlet("/profile-pic-upload-url")
+public class ProfilePicUploadUrlServlet extends HttpServlet {
 
   @Override
   public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-
-    UserService userService = UserServiceFactory.getUserService();
-    String urlToRedirectTo = "/account-creation-test.html";
-    String signUpUrl = userService.createLoginURL(urlToRedirectTo);
+    BlobstoreService blobstoreService = BlobstoreServiceFactory.getBlobstoreService();
+    String uploadUrl = blobstoreService.createUploadUrl("/user");
 
     response.setContentType("text/html");
-    response.getWriter().println(signUpUrl);
-    }
+    response.getWriter().println(uploadUrl);
   }
+}
