--- src/main/webapp/style.css
@@ -1,15 +1,0 @@
-.hidden {
-  display: none;
-}
-
-#profile-pic {
-  border-radius: 50%;
-  width: 75px;
-  height: 75px;
-  border:5px solid black;
-  line-height:50px;
-}
-
-.error-message {
-  color: red;
-}
